refactor: notification controller refactored
