+++ backend/src/controllers/notifications/index.ts
@@ -1,0 +1,2 @@
+export { default as addNotification } from './addNotification/addNotification';
+export { default as readNotifications } from './readNotifications/readNotifications';
