+++ backend/src/routing/userRouter.ts
@@ -3,7 +3,7 @@ import { Router } from 'express';
 import { USER_ROLES } from 'config/constants';
 import withAuth from 'middlewares/authMiddleware';
 import getProfileInformation from 'controllers/user/getProfileInfo';
-import readNotifications from 'controllers/notifications/readNotifications';
+import { readNotifications } from 'controllers/notifications';
 import { SubRoutes } from 'enums/routesEnum';
 
 const userRouter = Router();
